+++ R/excel.R
@@ -22,6 +22,29 @@
 #' @export
 #'
 #' @examples
+#' \dontrun{
+#'   # Create a simple matrix
+#'   mat <- matrix(1:6, nrow = 3, ncol = 2,
+#'                 dimnames = list(c("r1", "r2", "r3"),
+#'                                 c("c1", "c2")))
+#'   # Create a matsindf data frame
+#'   df <- tibble::tibble(mat = list(mat, mat, mat),
+#'                        worksheet_name = c("A", "B", "C"))
+#'   # Create a temporary file
+#'   mat_temp_path <- tempfile(pattern = "write_mat_to_excel_test_file",
+#'                             fileext = ".xlsx")
+#'   # Write the file.
+#'   df |>
+#'     write_mats_to_excel(mat_colname = "mat",
+#'                         worksheet_names = "worksheet_name",
+#'                         path = mat_temp_path,
+#'                         overwrite_file = TRUE)
+#'   # Check the appearance
+#'   openxlsx2::wb_open(mat_wb)
+#'   if (file.exists(mat_temp_path)) {
+#'     res <- file.remove(mat_temp_path)
+#'   }
+#' }
 write_mats_to_excel <- function(.psut_data = NULL,
                                 mat_colname,
                                 path,
@@ -166,16 +189,15 @@ write_mats_to_excel <- function(.psut_data = NULL,
   # Write the workbook
   mat_wb |>
     openxlsx2::wb_save(file = path, overwrite = overwrite_file)
-
 }
 
 
 #' Develop a warning message for malformed Excel worksheet names
 #'
-#' `write_ecc_to_excel()` can include worksheet names, but
+#' `write_mat_to_excel()` can include worksheet names, but
 #' it is important that they are legal names.
 #' This function emits a warning when `candidate_worksheet_names`
-#' is mal-formed.
+#' is malformed.
 #'
 #' @param candidate_worksheet_names Worksheet names to be checked.
 #'