+++ R/expand.R
@@ -55,8 +55,8 @@
 #'                                   "i1",  "i2",  "i1",  "i2", NA, NA),
 #'                   rowtypes  = c( ptype, ptype, ptype, ptype, ptype, itype, itype,
 #'                                  ptype, ptype, ptype, ptype, NA, NA),
-#'                   coltypes  = c(itype, itype, itype, itype, itype, ptype, ptype,
-#'                                 itype, itype, itype, itype, NA, NA),
+#'                   coltypes  = c( itype, itype, itype, itype, itype, ptype, ptype,
+#'                                  itype, itype, itype, itype, NA, NA),
 #'                   vals      = c(11  ,  22,    11 ,   22 ,   23 ,   11 ,   22 ,
 #'                                 11 ,   12 ,   11 ,   22,   0.2, 0.3)) %>%
 #'   group_by(Country, Year, matrix)