+++ NEWS.md
@@ -8,6 +8,7 @@ Cite all releases with doi [10.5281/zenodo.5475443](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
+## matsindf 0.4.11 (2025-11-12) 
 
 * New function `write_mats_to_excel()` writes
   matrices in a column of a `matsindf` data frame